requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision>=1.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
mlflow>=2.8.0
azureml-core>=1.54.0
azureml-mlflow>=1.54.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0